Complete saturation filter and re-organized the class struct

+++ src/include/f_edge_matrix.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Name            : edge_matrix.h
+ * Name            : f_edge_matrix.h
  * Project         : FlashPhoto
  * Module          : filter_manager
  * Description     : Header for EdgeMatrix class
@@ -38,7 +38,7 @@ class EdgeMatrix : public FilterMatrix {
    *
    * @param[in] incoming_size the new size of this matrix
    */
-  void Resize(float incoming_size, int detect_type);
+  void Resize(float incoming_size, float detect_type);
 
  private:
   int detect_type_;
